requirements.txt
requests==2.31.0
httpx==0.26.0
python-dotenv==1.0.1 